pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "databridge-core"
version = "1.2.0"
description = "Upload your Chart of Accounts. Get a production-ready financial hierarchy and dbt models. Zero config."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DataBridge AI", email = "hello@databridgeai.com" },
]
keywords = ["data", "reconciliation", "profiling", "etl", "finance", "csv", "diff", "fuzzy-match", "triage", "excel", "slack", "templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "pandas>=2.0",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
fuzzy = ["rapidfuzz>=3.0"]
pdf = ["pypdf>=3.0"]
ocr = ["pytesseract>=0.3", "Pillow>=9.0"]
sql = ["sqlalchemy>=2.0"]
triage = ["openpyxl>=3.1"]
all = [
    "rapidfuzz>=3.0",
    "pypdf>=3.0",
    "pytesseract>=0.3",
    "Pillow>=9.0",
    "sqlalchemy>=2.0",
    "openpyxl>=3.1",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "build>=1.0",
]

[project.urls]
Homepage = "https://github.com/datanexum/databridge-core"
Documentation = "https://github.com/datanexum/databridge-core#readme"
Repository = "https://github.com/datanexum/databridge-core"
Issues = "https://github.com/datanexum/databridge-core/issues"

[project.scripts]
databridge = "databridge_core.cli:cli"
databridge-tour = "databridge_core.examples.guided_tour:run_tour"

[tool.hatch.build.targets.wheel]
packages = ["src/databridge_core"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
